addressing issues 16 and 33

+++ v2 to v3/ETL Scripts/cz_omop_pcornet_concept_map_ddl.sql
@@ -3,7 +3,7 @@
   target_concept character varying(200),
   source_concept_class character varying(200),
   source_concept_id character varying(200),
-  value_as_concept_id integer,
+  value_as_concept_id bigint, 
   concept_description character varying(200)
 )
